Weitere Änderungen nach Versuchen im Keller
1) "nach Erreichen dieses Wegpunktes, Pfad vollständig ausgeführt."
2) "nächster Wegpunkt"
3) ließmich aktualisiert um Hysteresen
4) Warnung, dass Pfadsteuerung nicht beginnen kann, wenn nicht alle
Achsen posOK
5) Problem mit Integer bei Tabelle -> Pfad Funktion gelöst. Ist jetzt
.toFloat

+++ csvHandling.cpp
@@ -9,7 +9,7 @@ void MainWindow::on_acceptPath_clicked()
             for (int x = 0; x < 8; ++x) {
                 index = model->index(i,x);
                 if(index.isValid())
-                {   path[x][i] = model->data(index).toInt();
+                {   path[x][i] = model->data(index).toFloat();
                 }
                 else qDebug() << "invalid String";
             }
